requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
mock==1.0.1
webtest==2.0.16
factory-boy==2.4.0

# Management script
Flask-Script
